36: valid sudoku [M]
update the 2d and 3d array initialization, use {0} directly rather than
memset.
check if the sudoku is valid, notice: the index in this code represents
which number it is.

that is: if we find 5 in first row, then row[0][4] will be 1, we
traverse every char in the whole sudoku, check if there are any element
that are more than 1, we return false, or we return true.