update: Backtrack, Binary tree, D&C, DP, LRU&LFU
learn the backtrack, use the lambda function for DFS in the backtrack
function. this is trick way to write, but not hard to understand.

The TIME COMPLEXITY of NQUEEN is O(n^2 * n!), space is O(N).

Learn the D&C solution to solve the binary tree of all possilbe Full
binary trees.

Also learn how to write the KMP, the prefix array of KMP.

revise the LRU and LFU, the lfu is a bit hard.

next to revise the BINARY TREE and SOME DP problem, also don't forget
the sliding window problem.